% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csse_data.R
\docType{class}
\name{csse_data}
\alias{csse_data}
\title{CSSE data storage and handling class
An R6 class allowing to store, manipulate and retrieve CSSE data in a systematized way}
\format{
An \code{\link{R6Class}} generator object for CSSE data
}
\value{
Object of \code{\link{R6Class}} with methods for manipulate CSSE data
}
\description{
CSSE data storage and handling class
An R6 class allowing to store, manipulate and retrieve CSSE data in a systematized way

CSSE data storage and handling class
An R6 class allowing to store, manipulate and retrieve CSSE data in a systematized way
}
\keyword{CSSE}
\keyword{class}
\keyword{data}
\keyword{function}
\keyword{initialize}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{csse_data$new()}}
\item \href{#method-get_data}{\code{csse_data$get_data()}}
\item \href{#method-doubling_time}{\code{csse_data$doubling_time()}}
\item \href{#method-rolling_data}{\code{csse_data$rolling_data()}}
\item \href{#method-main_indicators}{\code{csse_data$main_indicators()}}
\item \href{#method-clone}{\code{csse_data$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{csse_data$new(case_type, iso3c_country_code, case_date, case_value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{case_type}}{a character string. Possible values : "confirmed", "deaths", "recovered"}

\item{\code{iso3c_country_code}}{a vector of iso3c codes}

\item{\code{case_date}}{a vector of dates}

\item{\code{case_value}}{a table with values by iso3c code (rows) and dates (cols)
data retrieval function}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-get_data}{}}}
\subsection{Method \code{get_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{csse_data$get_data(
  iso3c_country_code = NA,
  start_date = NA,
  end_date = NA,
  data_type = "cumulative",
  by_p100000_incidence = FALSE,
  add_suppl_info = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{iso3c_country_code}}{a vector of iso3c codes}

\item{\code{start_date}}{start date of data}

\item{\code{end_date}}{end date of data}

\item{\code{data_type}}{data type : "cumulative" or "daily"}

\item{\code{by_p100000_incidence}}{a logical value}

\item{\code{add_suppl_info}}{a logical value: indicates whether the results are reported in frequency or incidence
Function for calculating the doubling time of cases (at a given date)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-doubling_time"></a>}}
\if{latex}{\out{\hypertarget{method-doubling_time}{}}}
\subsection{Method \code{doubling_time()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{csse_data$doubling_time(iso3c_country_code = NA, n_days = 7, min_cases = 100)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{iso3c_country_code}}{a vector of iso3c codes}

\item{\code{n_days}}{the number of days used to calculate the growth rate}

\item{\code{min_cases}}{minimum number of cases to calculate the doubling time
Sum or moving average function}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rolling_data"></a>}}
\if{latex}{\out{\hypertarget{method-rolling_data}{}}}
\subsection{Method \code{rolling_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{csse_data$rolling_data(
  iso3c_country_code = NA,
  n_days = 7,
  rolling_function = "mean",
  by_p100000_incidence = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{iso3c_country_code}}{a vector of iso3c codes}

\item{\code{n_days}}{size of the sliding window in days}

\item{\code{rolling_function}}{a character string indicating whether the average or the sum is to be calculated}

\item{\code{by_p100000_incidence}}{a logical value
Extract de main epidemiological indicators}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-main_indicators"></a>}}
\if{latex}{\out{\hypertarget{method-main_indicators}{}}}
\subsection{Method \code{main_indicators()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{csse_data$main_indicators(iso3c_country_code = NA, as_of = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{iso3c_country_code}}{a vector of iso3c codes}

\item{\code{as_of}}{a cut-off date}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a table with the main epidemiological indicators
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{csse_data$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
