% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extended_countrycode.R
\name{extended_countrycode}
\alias{extended_countrycode}
\title{Extension of the countrycode function
A function to manage a wider range of iso3c codes and country names in French}
\usage{
extended_countrycode(x, origin, destination)
}
\arguments{
\item{x}{Vector which contains the codes or country names to be converted}

\item{origin}{Coding scheme of origin (string such as "iso3c" enclosed in quotes "")}

\item{destination}{Coding scheme of destination (string such as "iso3c" enclosed in quotes "")}
}
\value{
x values in the destination coding scheme
}
\description{
Extension of the countrycode function
A function to manage a wider range of iso3c codes and country names in French
}
