% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_through_system_proxy.R
\name{get_through_system_proxy}
\alias{get_through_system_proxy}
\title{Uploads a document through the system's transparent proxy
A function to retrieve the proxy settings from the system and pass them as parameters to the GET function of the httr package.}
\usage{
get_through_system_proxy(target_url)
}
\arguments{
\item{target_url}{a character string}
}
\value{
a list containing the results of the HTTP/HTTPS request
}
\description{
Uploads a document through the system's transparent proxy
A function to retrieve the proxy settings from the system and pass them as parameters to the GET function of the httr package.
}
