% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_in.R
\name{\%not_in\%}
\alias{\%not_in\%}
\title{Not in
Checks that value of x is not in the table vector.}
\usage{
x \%not_in\% table
}
\arguments{
\item{x}{vector or NULL: the values not to be matched. Long vectors are supported.}

\item{table}{vector or NULL: the values to be matched against. Long vectors are not supported.}
}
\value{
a logical vector
}
\description{
Not in
Checks that value of x is not in the table vector.
}
