% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_csse_data.R
\name{load_csse_data}
\alias{load_csse_data}
\title{Function for loading CSSE COVID-19 data from github}
\usage{
load_csse_data(
  case_type = "confirmed",
  sum_by_iso3c = TRUE,
  use_french_gouv_data = TRUE
)
}
\arguments{
\item{case_type}{a character string. Possible values : "confirmed", "deaths", "recovered"}

\item{sum_by_iso3c}{a logical variable, indicating whether cases should be aggregated by iso3c code}

\item{use_french_gouv_data}{a logical value}
}
\value{
a dataframe with the last csse github data
}
\description{
Function for loading CSSE COVID-19 data from github
}
