% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_oms_data.R
\name{load_oms_data}
\alias{load_oms_data}
\title{WHO COVID Sitrep data extractor}
\usage{
load_oms_data()
}
\value{
a table with the transmission class for each country
}
\description{
WHO COVID Sitrep data extractor
}
