% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rccs_dygraph_inc_plot.R
\name{rccs_dygraph_inc_plot}
\alias{rccs_dygraph_inc_plot}
\title{Dygraphs time series graph
A function to generate dynamic time series graphs}
\usage{
rccs_dygraph_inc_plot(
  csse_data_object,
  iso3c_country_code,
  by_p100000_incidence = FALSE,
  plot_title = NA,
  roll_period = 1,
  y_logscale = FALSE,
  data_type = "daily"
)
}
\arguments{
\item{csse_data_object}{a CsseData object created by load_csse}

\item{iso3c_country_code}{a character vector of iso3c country codes}

\item{by_p100000_incidence}{a logical value}

\item{plot_title}{a character string containing the plot title}

\item{roll_period}{the rolling period (smoothing)}

\item{y_logscale}{a logical value}

\item{data_type}{data type : "cumulative" or "daily"}
}
\value{
a dygraph object
}
\description{
Dygraphs time series graph
A function to generate dynamic time series graphs
}
