% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rccs_ggplot_doubling_time.R
\name{rccs_ggplot_doubling_time}
\alias{rccs_ggplot_doubling_time}
\title{GGplot2 doubling time graph
A function to generate static doubling time graphs}
\usage{
rccs_ggplot_doubling_time(
  csse_data_object,
  iso3c_country_code,
  plot_title = NA,
  window_size = 7,
  min_cases = 100,
  start_date = NA,
  end_date = NA
)
}
\arguments{
\item{csse_data_object}{a CsseData object created by load_csse}

\item{iso3c_country_code}{a character vector of iso3c country codes}

\item{plot_title}{a character string containing the plot title}

\item{window_size}{the size of the sliding window used to estimate the doubling time}

\item{min_cases}{the minimal number of cases to estimate the doubling time}

\item{start_date}{either NA or the start date of data}

\item{end_date}{either NA of the end date of data}
}
\value{
a ggplot2 object
}
\description{
GGplot2 doubling time graph
A function to generate static doubling time graphs
}
