% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rccs_ggplot_inc_plot.R
\name{rccs_ggplot_inc_plot}
\alias{rccs_ggplot_inc_plot}
\title{GGplot2 time series graph
A function to generate static time series graphs}
\usage{
rccs_ggplot_inc_plot(
  csse_data_object,
  iso3c_country_code,
  by_p100000_incidence = FALSE,
  plot_title = NA,
  y_logscale = FALSE,
  data_type = "daily",
  smoothing = TRUE,
  start_date = NA,
  end_date = NA,
  separate_graphs = FALSE
)
}
\arguments{
\item{csse_data_object}{a CsseData object created by load_csse}

\item{iso3c_country_code}{a character vector of iso3c country codes}

\item{by_p100000_incidence}{a logical value}

\item{plot_title}{a character string containing the plot title}

\item{y_logscale}{a logical value}

\item{data_type}{data type : "cumulative" or "daily"}

\item{smoothing}{a logical value}

\item{start_date}{either NA or the start date of data}

\item{end_date}{either NA of the end date of data}

\item{separate_graphs}{a logicial value}
}
\value{
a ggplot2 object
}
\description{
GGplot2 time series graph
A function to generate static time series graphs
}
