% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rccs_leaflet_inc_map.R
\name{rccs_leaflet_inc_map}
\alias{rccs_leaflet_inc_map}
\title{Cumulative incidence Leaflet map
A function to generate a dynamic world map}
\usage{
rccs_leaflet_inc_map(
  csse_data_object,
  n_days = 7,
  as_of = NA,
  custom_brks = NA
)
}
\arguments{
\item{csse_data_object}{a CsseData object created by load_csse}

\item{n_days}{cumulative incidence period length}

\item{as_of}{the cut-off date}

\item{custom_brks}{custom breaks for legend intervals. Must contains lower an upper bounds.}
}
\value{
a leaflet object
}
\description{
Cumulative incidence Leaflet map
A function to generate a dynamic world map
}
