% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_html_through_system_proxy.R
\name{read_html_through_system_proxy}
\alias{read_html_through_system_proxy}
\title{Parse an HTML document through the system's transparent proxy
A function downloading a document through the transparent proxy of the system then parsing it using function read_html of the xml2 package}
\usage{
read_html_through_system_proxy(target_url)
}
\arguments{
\item{target_url}{a character string}
}
\value{
a xml_document object
}
\description{
Parse an HTML document through the system's transparent proxy
A function downloading a document through the transparent proxy of the system then parsing it using function read_html of the xml2 package
}
