% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retry.R
\name{retry}
\alias{retry}
\title{Retry a function in cas of failure
A function to restart a function in case of failure (source StackOverflow)}
\usage{
retry(
  expr,
  isError = function(x) "try-error" \%in\% class(x),
  maxErrors = 5,
  sleep = 0
)
}
\arguments{
\item{expr}{expression to be evaluated}

\item{isError}{error control function}

\item{maxErrors}{Maximum number of possible failures}

\item{sleep}{time in seconds between two attempts}
}
\value{
the results of expr
}
\description{
Retry a function in cas of failure
A function to restart a function in case of failure (source StackOverflow)
}
